+++ typescript/lib/util.ts
@@ -10,17 +10,24 @@ export const TYPE = {
 export interface IHttpEventResponse {
     statusCode?: number;
     body?: string;
+    headers?: { [key: string]: string };
+
     [key: string]: any;
 }
 
 // tslint:disable-next-line
-export interface IService {}
+export interface IService {
+    preHandle?: (event: any, context: any) => Promise<void>;
+    postHandle?: (event: any, context: any, resp: any) => Promise<any>;
+}
 
 // tslint:disable-next-line
-export interface IEventHandler {}
+export interface IEventHandler {
+}
 
 // @todo probably a better way to handle this
 let c: Container;
+
 export function register(container: Container): void {
     c = container;
 
@@ -33,6 +40,7 @@ export function register(container: Container): void {
         });
     });
 }
+
 export function getContainer(): Container {
     return c;
 }