+++ typescript/test/lib/script/template/handler.template.spec.ts
@@ -10,14 +10,19 @@ import {MetadataKey} from "../../../../lib/service";
 
 describe("Handler Template", () => {
 
+    const serviceName: string = "serviceName";
     const methodName: string = "methodName";
     const handlerName: string = "handlerName";
     const ContainerMock: TypeMoq.IMock<Container> = TypeMoq.Mock.ofType<Container>(Container);
+    const service: any = {};
     const handler: any = {};
 
     let template: any;
 
     beforeEach(() => {
+        ContainerMock
+            .setup((x) => x.getNamed(TypeMoq.It.isAny(), serviceName))
+            .returns((type: any, target: any) => service);
         ContainerMock
             .setup((x) => x.getNamed(TypeMoq.It.isAny(), handlerName))
             .returns((type: any, target: any) => handler);
@@ -32,6 +37,7 @@ describe("Handler Template", () => {
             }
         });
 
+        Reflect.defineMetadata("service", [], service.constructor);
         Reflect.defineMetadata("event_handler", [], handler.constructor);
         Reflect.defineMetadata("param", [], handler.constructor);
     });
@@ -49,7 +55,7 @@ describe("Handler Template", () => {
         };
         handler[methodName] = () => resp;
 
-        template.handle(methodName, handlerName, null, null, (err, val) => {
+        template.handle(serviceName, methodName, handlerName, null, null, (err, val) => {
             chai.expect(resp).to.deep.equal(val);
         });
     });
@@ -59,7 +65,7 @@ describe("Handler Template", () => {
             throw new Error("Hello world!");
         };
 
-        template.handle(methodName, handlerName, null, null, (err, val) => {
+        template.handle(serviceName, methodName, handlerName, null, null, (err, val) => {
             chai.expect(err).to.not.be.undefined; // tslint:disable-line
             chai.expect(err).to.not.be.null; // tslint:disable-line
             chai.expect(err.message).to.equal("Hello world!");
@@ -74,7 +80,7 @@ describe("Handler Template", () => {
             throw error;
         };
 
-        template.handle(methodName, handlerName, null, null, (err, val) => {
+        template.handle(serviceName, methodName, handlerName, null, null, (err, val) => {
             chai.expect(val).to.not.be.undefined; // tslint:disable-line
             chai.expect(val).to.not.be.null; // tslint:disable-line
             chai.expect(err).to.be.null; // tslint:disable-line
@@ -129,7 +135,7 @@ describe("Handler Template", () => {
 
         Reflect.defineMetadata("event_handler", [metadata, metadataTwo], handler.constructor);
 
-        await template.handle(methodName, handlerName, event, context, (err, val) => {
+        await template.handle(serviceName, methodName, handlerName, event, context, (err, val) => {
             chai.expect(err).to.be.null; // tslint:disable-line
             chai.expect(val).to.be.undefined; // tslint:disable-line
         });
@@ -191,7 +197,7 @@ describe("Handler Template", () => {
                 name: key
             });
 
-            template.handle(methodName, handlerName, event, context, (err, val) => {
+            template.handle(serviceName, methodName, handlerName, event, context, (err, val) => {
                 chai.expect(err).to.be.null; // tslint:disable-line
                 chai.expect(val).to.be.undefined; // tslint:disable-line
             });
@@ -229,7 +235,7 @@ describe("Handler Template", () => {
                 type: "context"
             });
 
-            template.handle(methodName, handlerName, event, context, (err, val) => {
+            template.handle(serviceName, methodName, handlerName, event, context, (err, val) => {
                 chai.expect(err).to.be.null; // tslint:disable-line
                 chai.expect(val).to.be.undefined; // tslint:disable-line
             });
@@ -245,7 +251,7 @@ describe("Handler Template", () => {
                 name: "testParamKey"
             });
 
-            template.handle("", handlerName, event, context, (err, val) => {
+            template.handle(serviceName, "", handlerName, event, context, (err, val) => {
                 chai.expect(err.message).to.equal("Method for event handler not found!"); // tslint:disable-line
                 chai.expect(val).to.be.undefined; // tslint:disable-line
             });
@@ -280,7 +286,7 @@ describe("Handler Template", () => {
                 type: "context"
             });
 
-            template.handle(methodName, handlerName, event, context, (err, val) => {
+            template.handle(serviceName, methodName, handlerName, event, context, (err, val) => {
                 chai.expect(err).to.be.null; // tslint:disable-line
                 chai.expect(val).to.be.undefined; // tslint:disable-line
             });
@@ -458,7 +464,7 @@ describe("Handler Template", () => {
                 return resp;
             };
 
-            template.handle(methodName, handlerName, null, null, (err, val) => {
+            template.handle(serviceName, methodName, handlerName, null, null, (err, val) => {
                 chai.expect(resp).to.deep.equal(val);
             });
         });
@@ -506,7 +512,7 @@ describe("Handler Template", () => {
                 .setup((x) => x.getNamed(TypeMoq.It.isAny(), customHandlerName))
                 .returns((type: any, target: any) => customHandler);
 
-            template.handle(methodName, customHandlerName, methodName, {}, (err, res) => {
+            template.handle(serviceName, methodName, customHandlerName, methodName, {}, (err, res) => {
                 chai.expect(err.message).to.equal("Error thrown in handler function");
                 chai.expect(res).to.be.undefined; // tslint:disable-line
             });
@@ -537,7 +543,7 @@ describe("Handler Template", () => {
                 .setup((x) => x.getNamed(TypeMoq.It.isAny(), customHandlerName))
                 .returns((type: any, target: any) => customHandler);
 
-            template.handle(methodName, customHandlerName, methodName, {}, (err, res) => {
+            template.handle(serviceName, methodName, customHandlerName, methodName, {}, (err, res) => {
                 chai.expect(err).to.be.null; // tslint:disable-line
                 chai.expect(res).to.equal(customResp);
             });