+++ typescript/test/mock/index.ts
@@ -4,6 +4,6 @@ import {MockService} from "./mock-service";
 
 const container: Container = new Container();
 
-container.bind<IService>(TYPE.Service).to(MockService);
+container.bind<IService>(TYPE.Service).to(MockService).whenTargetNamed("MockService");
 
 register(container);